tut_LC/requirements.txt
langchain
langchain-core
langchain-huggingface
huggingface-hub
transformers
python-dotenv
numpy
scikit-learn